Rename:  폴더 이동
- 디자인 모델 의미에 맞게 Layout.tsx Template -> atoms로 이동
