+++ src/components/CalenderWrapper.tsx
@@ -1,6 +1,6 @@
 import DateSelectorBar from "@components/organisms/DateSelectorBar";
 import { ReactNode } from "react";
-import { CenteredLayoutComponent } from "@components/template/Layout";
+import { CenteredLayoutComponent } from "@components/atoms/Layout";
 
 /* 년도 월일 선택 바
  * {childeren}
