+++ src/components/mocules/navBar/HomeNavBar.tsx
@@ -1,6 +1,6 @@
 import TypeSelector from "@components/mocules/TypeSelector";
 import AllOrRecommendationToggleBar from "@components/mocules/ToggleBar";
-import { SpaceBetweenNavBar } from "@components/template/Layout";
+import { SpaceBetweenNavBar } from "@components/atoms/Layout";
 
 export function HomeNavBar() {
   return (
